+++ src/FSharpVSPowerTools/UnionMatchCaseGeneratorSmartTaggerProvider.cs
@@ -25,6 +25,9 @@ namespace FSharpVSPowerTools
         [Import]
         private ITextUndoHistoryRegistry undoHistoryRegistry = null;
 
+        [Import(typeof(ProjectFactory))]
+        private ProjectFactory projectFactory = null;
+
         public ITagger<T> CreateTagger<T>(ITextView textView, ITextBuffer buffer) where T : ITag
         {
             // Only provide the smart tagger on the top-level buffer
@@ -36,7 +39,7 @@ namespace FSharpVSPowerTools
             {
                 return new UnionMatchCaseGeneratorSmartTagger(textView, buffer,
                     undoHistoryRegistry.RegisterHistory(buffer),
-                    fsharpVsLanguageService, serviceProvider) as ITagger<T>;
+                    fsharpVsLanguageService, serviceProvider, projectFactory) as ITagger<T>;
             }
             else
                 return null;