+++ src/FSharpVSPowerTools/GeneralOptionsPage.cs
@@ -35,7 +35,7 @@ namespace FSharpVSPowerTools
             InterfaceImplementationEnabled = true;
             FolderOrganizationEnabled = false;
             FindAllReferencesEnabled = true;
-            RecordStubGenerationEnabled = true;
+            GenerateRecordStubEnabled = true;
             UnionPatternMatchCaseGenerationEnabled = true;
         }
 
@@ -137,7 +137,7 @@ namespace FSharpVSPowerTools
         public bool FindAllReferencesEnabled { get; set; }
 
         [DesignerSerializationVisibility(DesignerSerializationVisibility.Visible)]
-        public bool RecordStubGenerationEnabled { get; set; }
+        public bool GenerateRecordStubEnabled { get; set; }
 
         [DesignerSerializationVisibility(DesignerSerializationVisibility.Visible)]
         public bool UnionPatternMatchCaseGenerationEnabled { get; set; }
@@ -147,24 +147,7 @@ namespace FSharpVSPowerTools
         {
             get
             {
-                _optionsControl = new GeneralOptionsControl
-                                  {
-                                      OptionsPage = this,
-                                      XmlDocEnabled = this.XmlDocEnabled,
-                                      FormattingEnabled = this.FormattingEnabled,
-                                      NavBarEnabled = this.NavBarEnabled && this._navBarEnabledInAppConfig,
-                                      HighlightUsageEnabled = this.HighlightUsageEnabled,
-                                      RenameRefactoringEnabled = this.RenameRefactoringEnabled,
-                                      DepthColorizerEnabled = this.DepthColorizerEnabled,
-                                      NavigateToEnabled = this.NavigateToEnabled,
-                                      SyntaxColoringEnabled = this.SyntaxColoringEnabled,
-                                      InterfaceImplementationEnabled = this.InterfaceImplementationEnabled,
-                                      FolderOrganizationEnabled = this.FolderOrganizationEnabled,
-                                      FindAllReferencesEnabled = this.FindAllReferencesEnabled,
-                                      RecordStubGenerationEnabled = this.RecordStubGenerationEnabled,
-                                      UnionPatternMatchCaseGenerationEnabled = this.UnionPatternMatchCaseGenerationEnabled
-                                  };
-
+                _optionsControl = new GeneralOptionsControl(this);
                 return _optionsControl;
             }
         }
@@ -192,7 +175,7 @@ namespace FSharpVSPowerTools
                 InterfaceImplementationEnabled = _optionsControl.InterfaceImplementationEnabled;
                 FolderOrganizationEnabled = _optionsControl.FolderOrganizationEnabled;
                 FindAllReferencesEnabled = _optionsControl.FindAllReferencesEnabled;
-                RecordStubGenerationEnabled = _optionsControl.RecordStubGenerationEnabled;
+                GenerateRecordStubEnabled = _optionsControl.GenerateRecordStubEnabled;
                 UnionPatternMatchCaseGenerationEnabled = _optionsControl.UnionPatternMatchCaseGenerationEnabled;
             }